# GenAIFaceRegeneration-fyp/GenAIFaceRegeneration-server/openapi.yaml
openapi: 3.0.3
info:
  title: Pencil2Pixel API
  description: |
    Convert pencil sketches into realistic images using AI-powered UNet Generator model.
    
    The API supports user authentication, single and batch image processing with customizable 
    attributes, and maintains a history of generated images for each user.
    
    **Authentication:** Most endpoints require JWT token authentication.
  version: 2.1.0
  contact:
    name: API Support
    email: info.adnansultan@gmail.com
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: http://0.0.0.0:5000
    description: Local server (all interfaces)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from signup or login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message description
      required:
        - error
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
      required:
        - user_id
        - username
        - email
...
security:
  - BearerAuth: []
paths:
...
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: User Signup
      description: Create a new user account
      operationId: signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  minLength: 3
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
            example:
              username: john_doe
              email: john@example.com
              password: securepass123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Login to existing account
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              email: john@example.com
              password: securepass123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get Profile
      description: Get current user profile information
      operationId: getProfile
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    allOf:
                      - $ref: '#/components/schemas/User'
                      - type: object
                        properties:
                          created_at:
                            type: string
                            format: date-time
                          total_images:
                            type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Authentication
      summary: Update Profile
      description: Update user profile information
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
            example:
              username: new_username
              email: newemail@example.com
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /generate:
    post:
      tags:
        - Generation
      summary: Generate Single Image
      description: |
        Convert a single pencil sketch to a realistic image using the UNet Generator model.
        
        Upload an image file and optionally provide 20 comma-separated integer attribute values
        via the `attributes` parameter. Attributes are retained for API compatibility but have
        no measurable effect on the generated output due to InstanceNorm in the model architecture.
      operationId: generateImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: Pencil sketch image file (PNG, JPG, etc.)
                attributes:
                  type: string
                  pattern: '^(-?\d+,){19}-?\d+$'
                  example: "1,0,1,0,1,0,0,1,1,0,0,1,0,1,0,0,1,1,0,0"
                  description: 20 comma-separated integer values (retained for compatibility, no effect on output)
                format:
                  type: string
                  enum: [image, base64]
                  default: image
                  description: Response format - 'image' returns PNG file, 'base64' returns JSON
                save:
                  type: string
                  enum: ['true', 'false']
                  default: 'false'
                  description: Set to 'true' to save image to history
                quality:
                  type: string
                  enum: [low, medium, high, ultra]
                  default: medium
                  description: Quality preset for post-processing
                upscale:
                  type: string
                  enum: ['true', 'false']
                  default: 'false'
                  description: Set to 'true' to upscale output to 512x512
            examples:
              minimal:
                summary: Generate with defaults
                value:
                  image: <binary>
                  format: base64
                  save: 'true'
              withAttributes:
                summary: Generate with raw attribute values
                value:
                  image: <binary>
                  attributes: "1,0,1,0,1,0,0,1,1,0,0,1,0,1,0,0,1,1,0,0"
                  format: image
      responses:
        '200':
          description: Image generated successfully
          content:
            image/png:
              schema:
                type: string
                format: binary
                description: Generated PNG image (when format=image)
            application/json:
              schema:
                type: object
                properties:
                  image:
                    type: string
                    description: Base64-encoded PNG image
                  format:
                    type: string
                    example: base64
                  quality:
                    type: string
                    example: medium
                  size:
                    type: array
                    items:
                      type: integer
                    example: [256, 256]
                  image_id:
                    type: string
                    format: uuid
                    description: Image ID (only when save=true)
                  saved:
                    type: boolean
                    description: Whether image was saved (only when save=true)
              examples:
                base64Response:
                  value:
                    image: "iVBORw0KGgoAAAANSUhEUgAAAAE..."
                    format: base64
                    quality: medium
                    size: [256, 256]
        '400':
          description: Bad request - missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                noImage:
                  value:
                    error: "No image file provided"
                emptyFilename:
                  value:
                    error: "Empty filename"
                invalidAttributes:
                  value:
                    error: "Attributes must be 20 values"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error during processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...